Overwrite with env variable

+++ packages/babel-parser/test/helpers/runFixtureTests.js
@@ -131,12 +131,14 @@ function runTest(test, parseFunction) {
       if (err.message === opts.throws) {
         return;
       } else {
-        /*
-        const fn = path.dirname(test.expect.loc) + "/options.json";
-        test.options = test.options || {};
-        test.options.throws = err.message;
-        fs.writeFileSync(fn, JSON.stringify(test.options, null, "  "));
-        */
+        if (process.env.OVERWRITE) {
+          const fn = path.dirname(test.expect.loc) + "/options.json";
+          test.options = test.options || {};
+          test.options.throws = err.message;
+          fs.writeFileSync(fn, JSON.stringify(test.options, null, "  "));
+          return;
+        }
+
         err.message =
           "Expected error message: " +
           opts.throws +
